+++ src/sections/Epilogue.vue
@@ -31,7 +31,13 @@ let timeDisplay: ComputedRef<string> = computed(() => {
 
 <template>
   <SectionBorder border-colour-class="border-sky-500">
-    <SectionTitle version="3" :stability="Stability.Stable" anchor="epilogue" title="Meta" />
+    <SectionTitle
+      version="3.1"
+      :stability="Stability.Stable"
+      anchor="epilogue"
+      title="Meta"
+      :add-toc="false"
+    />
     <div class="flex">
       <div class="flex-1 space-y-2">
         <Link :href="getSetting('app_gitlab_url')">GitLab Project</Link>