+++ src/sections/Ideas.vue
@@ -13,7 +13,7 @@ const feedbackText =
 
 <template>
   <SectionBorder border-colour-class="border-sky-500">
-    <SectionTitle version="2.0" anchor="ideas" title="Ideas" />
+    <SectionTitle version="2.1" anchor="ideas" title="Ideas" :add-toc="false" />
     <div class="space-y-4">
       <GuidanceText>{{ feedbackText }}</GuidanceText>
       <div class="space-y-2">